add AWS_REGION env variable

+++ aws/create_and_upload_self_signed_cert_acm.sh
@@ -1,15 +1,25 @@
 #!/bin/bash
 
+if [ -z $AWS_REGION ]
+then
+  echo "Missing AWS_REGION"
+  exit
+fi
+
 # if this SSM value exists, abort script
-CERT_EXISTS_ALREADY=$(aws ssm get-parameter --name /sublime-security/self-signed-acm-certificate-for-sublime-security-on-prem-deployment | jq .Parameter.Value)
+CERT_EXISTS_ALREADY=$(aws ssm get-parameter --region $AWS_REGION --name /sublime-security/self-signed-acm-certificate-for-sublime-security-on-prem-deployment | jq .Parameter.Value)
 
 if [ -z $CERT_EXISTS_ALREADY ]
 then
     echo "ACM and SSM parameter for ACM certificate need to be created"
     openssl genrsa 2048 > my-aws-private.key
+    echo "Private key generated"
     openssl req -new -x509 -nodes -sha1 -days 36500 -extensions v3_ca -key my-aws-private.key -subj "/C=US/ST=DC/L=Washington, DC/O=Sublime Security/OU=Sublime Security on-prem deployment/CN=sublime-security-self-hosted-platform.local/emailAddress=support@sublimesecurity.com/" > my-aws-public.crt
-    CERT_ARN="$(aws acm import-certificate --certificate fileb://my-aws-public.crt --private-key fileb://my-aws-private.key --tags '[{"Key":"Function", "Value": "Self-signed certificate for Sublime Security on-prem deployment"}]' | jq -r .CertificateArn)"
-    aws ssm put-parameter --name "/sublime-security/self-signed-acm-certificate-for-sublime-security-on-prem-deployment" --description "ARN for self-signed SSL certificate created for Sublime Security self-hosted deployment" --value $CERT_ARN --type "String" --tags '[{"Key":"Function","Value": "ARN for self-signed SSL certificate created for Sublime Security self-hosted deployment"}, {"Key":"Purpose", "Value": "Used for the Sublime Security self-hosted deployment"}, {"Key": "Date created", "Value":  "April 21 2021"}]'
+    echo "Public key generated"
+    CERT_ARN="$(aws acm import-certificate --certificate fileb://my-aws-public.crt --private-key fileb://my-aws-private.key --region $AWS_REGION --tags '[{"Key":"Function", "Value": "Self-signed certificate for Sublime Security on-prem deployment"}]' | jq -r .CertificateArn)"
+    echo "CERT generation status: $CERT_ARN"
+    OUTPUT=$(aws ssm put-parameter --name "/sublime-security/self-signed-acm-certificate-for-sublime-security-on-prem-deployment"  --region $AWS_REGION --description "ARN for self-signed SSL certificate created for Sublime Security self-hosted deployment" --value $CERT_ARN --type "String" --tags '[{"Key":"Function","Value": "ARN for self-signed SSL certificate created for Sublime Security self-hosted deployment"}, {"Key":"Purpose", "Value": "Used for the Sublime Security self-hosted deployment"}, {"Key": "Date created", "Value":  "April 21 2021"}]')
+    echo "SSM Parameter output: $OUTPUT"
     rm my-aws-private.key;
     rm my-aws-public.crt;
